Add BuildslaveName to HTCondor job ad information; set BuildslaveJob ClassAd
The job_ad_information_attr directive in a Condor submit causes "[t]he named
attributes and their values [to be] written to the job event log whenever
any event is being written to the log."

The BuildslaveJob is a boolean ClassAd set to TRUE for use in rank
and requirements expressions in the HTCondor config, akin to the
InteractiveJob ClassAd. This allows Buildbot jobs to be easily
identified so they can be steered toward or away from specific machines
in the pool.